requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --strip-extras
#
aiohttp==3.9.1
    # via
    #   langchain
    #   langchain-community
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.6.0
    # via pydantic
anyio==4.2.0
    # via
    #   httpx
    #   langchain-core
    #   openai
asgiref==3.7.2
    # via django
attrs==23.1.0
    # via aiohttp
certifi==2023.11.17
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via
    #   pdfminer-six
    #   requests
click==8.1.7
    # via uvicorn
cryptography==41.0.7
    # via pdfminer-six
dataclasses-json==0.6.3
    # via
    #   langchain
    #   langchain-community
distro==1.8.0
    # via openai
django==5.0
    # via -r requirements.in
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
google-search-results==2.4.2
    # via -r requirements.in
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.2
    # via httpx
httpx==0.25.2
    # via
    #   -r requirements.in
    #   openai
    #   python-telegram-bot
idna==3.6
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jsonpatch==1.33
    # via
    #   langchain
    #   langchain-core
jsonpointer==2.4
    # via jsonpatch
langchain==0.0.352
    # via -r requirements.in
langchain-community==0.0.6
    # via langchain
langchain-core==0.1.3
    # via
    #   langchain
    #   langchain-community
langchainhub==0.1.14
    # via -r requirements.in
langsmith==0.0.75
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
marshmallow==3.20.1
    # via dataclasses-json
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
numexpr==2.8.8
    # via -r requirements.in
numpy==1.26.2
    # via
    #   langchain
    #   langchain-community
    #   numexpr
openai==1.6.1
    # via -r requirements.in
packaging==23.2
    # via
    #   langchain-core
    #   marshmallow
pdfminer-six==20221105
    # via pdfplumber
pdfplumber==0.10.3
    # via -r requirements.in
pillow==10.1.0
    # via pdfplumber
promptlayer==0.4.0
    # via -r requirements.in
pycparser==2.21
    # via cffi
pydantic==2.5.3
    # via
    #   langchain
    #   langchain-core
    #   langsmith
    #   openai
pydantic-core==2.14.6
    # via pydantic
pypdf==3.17.3
    # via -r requirements.in
pypdfium2==4.25.0
    # via pdfplumber
python-dotenv==1.0.0
    # via -r requirements.in
python-telegram-bot==20.7
    # via -r requirements.in
pyyaml==6.0.1
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
requests==2.31.0
    # via
    #   google-search-results
    #   langchain
    #   langchain-community
    #   langchain-core
    #   langchainhub
    #   langsmith
    #   promptlayer
sniffio==1.3.0
    # via
    #   anyio
    #   httpx
    #   openai
sqlalchemy==2.0.23
    # via
    #   langchain
    #   langchain-community
sqlparse==0.4.4
    # via django
tenacity==8.2.3
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
tqdm==4.66.1
    # via openai
types-requests==2.31.0.10
    # via langchainhub
typing-extensions==4.9.0
    # via
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
urllib3==2.1.0
    # via
    #   requests
    #   types-requests
uvicorn==0.25.0
    # via -r requirements.in
yarl==1.9.4
    # via aiohttp
